# Reduce the iteration count to make the OSU tests run faster, especially on slower interconnects