Tests on e.g. the AWS Magic Castle cluster have been pretty slow, especially the collectives that send a ton of data around. A thousand iterations is definitely overkill, we don't generally need that kind of precision to see when something is wrong. 

An example run with `--reruns=5` and only the bandwidth test for one of the modules produced:

```
[----------] start processing checks
[ RUN      ] EESSI_OSU_Micro_Benchmarks_pt2pt %benchmark_info=mpi.pt2pt.osu_bw %scale=2_nodes %module_name=OSU-Micro-Benchmarks/7.2-gompi-2023b %device_type=cpu /f749ec9f @snellius:genoa+default
[       OK ] (1/1) EESSI_OSU_Micro_Benchmarks_pt2pt %benchmark_info=mpi.pt2pt.osu_bw %scale=2_nodes %module_name=OSU-Micro-Benchmarks/7.2-gompi-2023b %device_type=cpu /f749ec9f @snellius:genoa+default
P: bandwidth: 28178.61 MB/s (r:0, l:None, u:None)
[----------] all spawned checks have finished

[----------] start processing checks
[ RUN      ] EESSI_OSU_Micro_Benchmarks_pt2pt %benchmark_info=mpi.pt2pt.osu_bw %scale=2_nodes %module_name=OSU-Micro-Benchmarks/7.2-gompi-2023b %device_type=cpu /f749ec9f @snellius:genoa+default
[       OK ] (1/1) EESSI_OSU_Micro_Benchmarks_pt2pt %benchmark_info=mpi.pt2pt.osu_bw %scale=2_nodes %module_name=OSU-Micro-Benchmarks/7.2-gompi-2023b %device_type=cpu /f749ec9f @snellius:genoa+default
P: bandwidth: 28175.13 MB/s (r:0, l:None, u:None)
[----------] all spawned checks have finished

[----------] start processing checks
[ RUN      ] EESSI_OSU_Micro_Benchmarks_pt2pt %benchmark_info=mpi.pt2pt.osu_bw %scale=2_nodes %module_name=OSU-Micro-Benchmarks/7.2-gompi-2023b %device_type=cpu /f749ec9f @snellius:genoa+default
[       OK ] (1/1) EESSI_OSU_Micro_Benchmarks_pt2pt %benchmark_info=mpi.pt2pt.osu_bw %scale=2_nodes %module_name=OSU-Micro-Benchmarks/7.2-gompi-2023b %device_type=cpu /f749ec9f @snellius:genoa+default
P: bandwidth: 28173.91 MB/s (r:0, l:None, u:None)
[----------] all spawned checks have finished

[----------] start processing checks
[ RUN      ] EESSI_OSU_Micro_Benchmarks_pt2pt %benchmark_info=mpi.pt2pt.osu_bw %scale=2_nodes %module_name=OSU-Micro-Benchmarks/7.2-gompi-2023b %device_type=cpu /f749ec9f @snellius:genoa+default
[       OK ] (1/1) EESSI_OSU_Micro_Benchmarks_pt2pt %benchmark_info=mpi.pt2pt.osu_bw %scale=2_nodes %module_name=OSU-Micro-Benchmarks/7.2-gompi-2023b %device_type=cpu /f749ec9f @snellius:genoa+default
P: bandwidth: 28175.5 MB/s (r:0, l:None, u:None)
[----------] all spawned checks have finished

[----------] start processing checks
[ RUN      ] EESSI_OSU_Micro_Benchmarks_pt2pt %benchmark_info=mpi.pt2pt.osu_bw %scale=2_nodes %module_name=OSU-Micro-Benchmarks/7.2-gompi-2023b %device_type=cpu /f749ec9f @snellius:genoa+default
[       OK ] (1/1) EESSI_OSU_Micro_Benchmarks_pt2pt %benchmark_info=mpi.pt2pt.osu_bw %scale=2_nodes %module_name=OSU-Micro-Benchmarks/7.2-gompi-2023b %device_type=cpu /f749ec9f @snellius:genoa+default
P: bandwidth: 28170.4 MB/s (r:0, l:None, u:None)
[----------] all spawned checks have finished
```

That shows reproducibility is just fine with the current iteration count.

Sure, on interconnects with larger overbooking, you will probably see more variation - but that's a realistic representation of performance of a shared resource. Not necessarily a problem.